Refactor player to split responsibilities across classes
