+++ app/src/main/java/net/sigmabeta/chipbox/ui/main/MainPresenter.kt
@@ -3,7 +3,9 @@ package net.sigmabeta.chipbox.ui.main
 import android.media.session.PlaybackState
 import android.os.Bundle
 import net.sigmabeta.chipbox.R
-import net.sigmabeta.chipbox.backend.Player
+import net.sigmabeta.chipbox.backend.UiUpdater
+import net.sigmabeta.chipbox.backend.player.Player
+import net.sigmabeta.chipbox.backend.player.Playlist
 import net.sigmabeta.chipbox.model.domain.Game
 import net.sigmabeta.chipbox.model.events.GameEvent
 import net.sigmabeta.chipbox.model.events.PositionEvent
@@ -18,7 +20,9 @@ import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class MainPresenter @Inject constructor(val player: Player) : ActivityPresenter() {
+class MainPresenter @Inject constructor(val player: Player,
+                                        val playlist: Playlist,
+                                        val updater: UiUpdater) : ActivityPresenter() {
     var view: MainView? = null
 
     var state = player.state
@@ -44,9 +48,9 @@ class MainPresenter @Inject constructor(val player: Player) : ActivityPresenter(
         when (player.state) {
             PlaybackState.STATE_PLAYING -> player.pause()
 
-            PlaybackState.STATE_PAUSED -> player.play()
+            PlaybackState.STATE_PAUSED -> player.start(null)
 
-            PlaybackState.STATE_STOPPED -> player.play()
+            PlaybackState.STATE_STOPPED -> player.start(null)
         }
     }
 
@@ -63,7 +67,7 @@ class MainPresenter @Inject constructor(val player: Player) : ActivityPresenter(
     override fun updateViewState() {
         updateHelper()
 
-        val subscription = player.updater.asObservable()
+        val subscription = updater.asObservable()
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe {
                     when (it) {
@@ -95,11 +99,11 @@ class MainPresenter @Inject constructor(val player: Player) : ActivityPresenter(
     }
 
     private fun updateHelper() {
-        player.playingTrackId?.let {
+        playlist.playingTrackId?.let {
             displayTrack(it, false)
         }
 
-        player.playingGameId?.let {
+        playlist.playingGameId?.let {
             displayGame(it, true)
         }
