+++ app/src/main/java/net/sigmabeta/chipbox/ui/player/PlayerControlsPresenter.kt
@@ -3,7 +3,9 @@ package net.sigmabeta.chipbox.ui.player
 import android.media.session.PlaybackState
 import android.os.Bundle
 import net.sigmabeta.chipbox.R
-import net.sigmabeta.chipbox.backend.Player
+import net.sigmabeta.chipbox.backend.UiUpdater
+import net.sigmabeta.chipbox.backend.player.Player
+import net.sigmabeta.chipbox.backend.player.Playlist
 import net.sigmabeta.chipbox.dagger.scope.ActivityScoped
 import net.sigmabeta.chipbox.model.events.StateEvent
 import net.sigmabeta.chipbox.ui.BaseView
@@ -12,7 +14,9 @@ import rx.android.schedulers.AndroidSchedulers
 import javax.inject.Inject
 
 @ActivityScoped
-class PlayerControlsPresenter @Inject constructor(val player: Player) : FragmentPresenter() {
+class PlayerControlsPresenter @Inject constructor(val player: Player,
+                                                  val playlist: Playlist,
+                                                  val updater: UiUpdater) : FragmentPresenter() {
     var view: PlayerControlsView? = null
 
     var updatedOnce = false
@@ -49,7 +53,7 @@ class PlayerControlsPresenter @Inject constructor(val player: Player) : Fragment
      */
 
     override fun onReCreate(arguments: Bundle?, savedInstanceState: Bundle) {
-        if (player.playingTrackId == null && player.playbackQueue.isEmpty()) {
+        if (playlist.playingTrackId == null && playlist.playbackQueue.isEmpty()) {
             view?.finish()
         }
     }
@@ -65,7 +69,7 @@ class PlayerControlsPresenter @Inject constructor(val player: Player) : Fragment
     override fun updateViewState() {
         updateHelper()
 
-        val subscription = player.updater.asObservable()
+        val subscription = updater.asObservable()
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe {
                     when (it) {
@@ -95,22 +99,22 @@ class PlayerControlsPresenter @Inject constructor(val player: Player) : Fragment
     private fun onPlayPauseClick() {
         when (player.state) {
             PlaybackState.STATE_PLAYING -> player.pause()
-            PlaybackState.STATE_PAUSED -> player.play()
-            PlaybackState.STATE_STOPPED -> player.play()
+            PlaybackState.STATE_PAUSED -> player.start(null)
+            PlaybackState.STATE_STOPPED -> player.start(null)
         }
     }
 
     private fun toggleShuffle() {
-        player.shuffle = !player.shuffle
+        playlist.shuffle = !playlist.shuffle
 
         displayShuffle()
     }
 
     private fun toggleRepeat() {
-        player.repeat = if (player.repeat >= Player.REPEAT_ONE) {
+        playlist.repeat = if (playlist.repeat >= Player.REPEAT_ONE) {
             Player.REPEAT_OFF
         } else {
-            player.repeat + 1
+            playlist.repeat + 1
         }
 
         displayRepeat()
@@ -135,7 +139,7 @@ class PlayerControlsPresenter @Inject constructor(val player: Player) : Fragment
     }
 
     private fun displayShuffle() {
-        if (player.shuffle) {
+        if (playlist.shuffle) {
             view?.setShuffleEnabled()
         } else {
             view?.setShuffleDisabled()
@@ -143,7 +147,7 @@ class PlayerControlsPresenter @Inject constructor(val player: Player) : Fragment
     }
 
     private fun displayRepeat() {
-        when (player.repeat) {
+        when (playlist.repeat) {
             Player.REPEAT_OFF -> view?.setRepeatDisabled()
             Player.REPEAT_ALL -> view?.setRepeatAll()
             Player.REPEAT_ONE -> view?.setRepeatOne()
