+++ app/src/main/java/net/sigmabeta/chipbox/ui/track/TrackListPresenter.kt
@@ -2,7 +2,7 @@ package net.sigmabeta.chipbox.ui.track
 
 import android.os.Bundle
 import net.sigmabeta.chipbox.R
-import net.sigmabeta.chipbox.backend.Player
+import net.sigmabeta.chipbox.backend.player.Player
 import net.sigmabeta.chipbox.dagger.scope.ActivityScoped
 import net.sigmabeta.chipbox.model.domain.Artist
 import net.sigmabeta.chipbox.model.domain.Track
